introduce controller interface
This commit introduces a ControllerInterface interface which
every controller added to Kedge has to implement.

In the current code, we have only the Deployment controller and
hence the DeploymentSpecMod definition in Kedge spec has to
implement this interface.

The ControllerInterface interface has 4 methods inside it, namely
Unmarshal, Validate, Fix and Transform; and this commit refactors
such that every controller has a new file. Based on the
current code, this means that pkg/spec/deployment.go has been
added.

These files have the code pertaining to each method of the
interface each implements.

This separation based on files has removed the need for older
file structure, hence the files -
pkg/encoding/encoding.go
pkg/spec/fix.go
pkg/spec/kubernetes.go
pkg/spec/validate.go
have gone away and have been merged into deployment.go

Besides this, a lot of functions have been converted to methods
applying to the Controller they pertain to.

Tests have been fixed.

Please note that a lot more refactor can be done, but that should
happen in subsequent PRs.

Merging this refactor eases the way to put in other controllers
very easily, like Kubernetes Jobs which is lined up next.
