+++ pkg/spec/spec.go
@@ -17,6 +17,7 @@ limitations under the License.
 package spec
 
 import (
+	"fmt"
 	api_v1 "k8s.io/client-go/pkg/api/v1"
 	ext_v1beta1 "k8s.io/client-go/pkg/apis/extensions/v1beta1"
 )
@@ -96,16 +97,16 @@ type SecretMod struct {
 	api_v1.Secret `json:",inline"`
 }
 
-// App defines the complete kedge application
-type App struct {
+//ControllerFields are the common fields in every controller Kedge supports
+type ControllerFields struct {
 	// Name of the micro-service
 	Name string `json:"name"`
+	// Specify which Kubernetes controller to generate. Defaults to deployment.
+	Controller string `json:"controller,omitempty"`
 	// Map of string keys and values that can be used to organize and categorize
 	// (scope and select) objects. May match selectors of replication controllers
 	// and services. More info: http://kubernetes.io/docs/user-guide/labels
 	Labels map[string]string `json:"labels,omitempty"`
-	// Specify which type of Kubernetes controller to generate. Defaults to deployment.
-	Controller string `json:"controller,omitempty"`
 	// List of volume that should be mounted on the pod.
 	VolumeClaims []VolumeClaim `json:"volumeClaims,omitempty"`
 	// List of configMaps
@@ -119,10 +120,32 @@ type App struct {
 	// List of Kubernetes resource files, that can be directly given to Kubernetes
 	ExtraResources []string `json:"extraResources,omitempty"`
 
-	PodSpecMod                 `json:",inline"`
+	PodSpecMod `json:",inline"`
+}
+
+// DeploymentSpecMod is Kedge's extension of Kubernetes DeploymentSpec and allows
+// defining a complete kedge application
+type DeploymentSpecMod struct {
+	ControllerFields           `json:",inline"`
 	ext_v1beta1.DeploymentSpec `json:",inline"`
 }
 
 type Controller struct {
 	Controller string `json:"controller,omitempty"`
+}
+
+// Validate
+
+func validateVolumeClaims(vcs []VolumeClaim) error {
+	// find the duplicate volume claim names, if found any then error out
+	vcmap := make(map[string]interface{})
+	for _, vc := range vcs {
+		if _, ok := vcmap[vc.Name]; !ok {
+			// value here does not matter
+			vcmap[vc.Name] = nil
+		} else {
+			return fmt.Errorf("duplicate entry of volume claim %q", vc.Name)
+		}
+	}
+	return nil
 }
