--- pkg/spec/validate_test.go
@@ -1,32 +1,0 @@
-/*
-Copyright 2017 The Kedge Authors All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package spec
-
-import "testing"
-
-func TestValidateVolumeClaims(t *testing.T) {
-
-	failingTest := []VolumeClaim{{Name: "foo"}, {Name: "bar"}, {Name: "foo"}}
-
-	err := validateVolumeClaims(failingTest)
-	if err == nil {
-		t.Errorf("should have failed but passed for input: %+v", failingTest)
-	} else {
-		t.Logf("failed with error: %v", err)
-	}
-
-}
